+++ src/main/java/com/dao/IBaseDao.java
@@ -5,7 +5,7 @@ import java.util.List;
 
 public abstract interface IBaseDao<E> {
 	
-	public Long add(E entity);
+	public Integer add(E entity);
 
 	public void update(E entity);
 
@@ -13,8 +13,8 @@ public abstract interface IBaseDao<E> {
 	
 	public List<E> list(E entity);
 	
-	public E getEntityById(Long id);
+	public E getEntityById(Integer id);
 
-	public void remove(Long id);
+	public void remove(Integer id);
 
 }
